requirements.txt
chainlit
httpx
google-genai
